Clean up MessageTypes view
